requirements.txt
pandas==1.3.2
camelot-py[cv]==0.10.1
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23
python-dotenv>=0.5.1